Move fixture helpers into a seperate file
So they can be shared by the cukes too
